Cleanup client package, making it more organized and clean
