--- src/main/java/client/ui/dialogs/DialogContainer.java
@@ -1,13 +1,0 @@
-package client.ui.dialogs;
-
-import javafx.scene.Node;
-
-/**
- * @author Wesley Klop
- */
-public interface DialogContainer {
-
-    void showDialog(Node dialog);
-
-    void clearDialog();
-}
