+++ src/main/java/client/ui/dialogs/MessagesControllerProvider.java
@@ -1,4 +1,4 @@
-package client.ui;
+package client.ui.dialogs;
 
 public class MessagesControllerProvider {
