+++ src/main/java/client/ui/game/RouteCardStackLayoutController.java
@@ -1,5 +1,6 @@
-package client.ui;
+package client.ui.game;
 
+import client.ui.dialogs.MessagesControllerProvider;
 import javafx.fxml.FXML;
 
 public class RouteCardStackLayoutController {
