+++ src/main/java/client/ui/mainmenu/MainMenuLoadController.java
@@ -1,4 +1,4 @@
-package client.ui.MainMenuControllers;
+package client.ui.mainmenu;
 
 import javafx.scene.control.Label;
