+++ src/main/java/client/ui/mainmenu/MainMenuRulesController.java
@@ -1,4 +1,4 @@
-package client.ui.MainMenuControllers;
+package client.ui.mainmenu;
 
 import javafx.beans.binding.Bindings;
 import javafx.fxml.Initializable;
