+++ src/main/java/client/util/UserPreferences.java
@@ -1,4 +1,4 @@
-package client;
+package client.util;
 
 import util.Observer;
