+++ src/main/resources/views/layout_full_game.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.Pane?>
 <Pane xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.FullGameController"
+      fx:controller="client.ui.game.FullGameController"
       prefHeight="1080" prefWidth="1920"
       fx:id="rootPane" stylesheets="/css/defaultStyling.css" styleClass="gameScreen">
     <ImageView>
