+++ src/main/resources/views/layout_game_bank.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.VBox?>
 <VBox xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.LayoutBankController"
+      fx:controller="client.ui.game.LayoutBankController"
       prefHeight="780.0" prefWidth="200.0"
       spacing="10" fx:id="rootBox">
