+++ src/main/resources/views/layout_main_menu.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 <Pane xmlns:fx="http://javafx.com/fxml/1" prefHeight="1080.0" prefWidth="3840.0"
       xmlns="http://javafx.com/javafx/8.0.121"
-      fx:controller="client.ui.MainMenuControllers.MainMenuController" fx:id="rootPane"
+      fx:controller="client.ui.mainmenu.MainMenuController" fx:id="rootPane"
       stylesheets="/css/defaultStyling.css">
 
     <Pane fx:id="snowPane"/>
