+++ src/main/resources/views/layout_main_menu_options.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.VBox?>
 <VBox xmlns:fx="http://javafx.com/fxml" xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.MainMenuControllers.MainMenuOptionController" prefHeight="400.0" prefWidth="600.0"
+      fx:controller="client.ui.mainmenu.MainMenuOptionController" prefHeight="400.0" prefWidth="600.0"
       spacing="20">
     <ImageView fitWidth="550" preserveRatio="true">
         <Image url="/images/options.png"/>
