+++ src/main/resources/views/layout_main_menu_rules.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.text.Text?>
 <VBox xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.MainMenuControllers.MainMenuRulesController"
+      fx:controller="client.ui.mainmenu.MainMenuRulesController"
       prefHeight="400.0" prefWidth="600.0" spacing="5" alignment="CENTER">
 
     <ImageView fitWidth="550" preserveRatio="true">
