+++ src/main/resources/views/layout_route_card_stack.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.Pane?>
 <Pane xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.RouteCardStackLayoutController"
+      fx:controller="client.ui.game.RouteCardStackLayoutController"
       prefHeight="400.0" prefWidth="600.0">
 
     <ImageView onMouseClicked="#onMouseClickedAction">
