+++ src/main/resources/views/layout_routes_map.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.Pane?>
 <Pane xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.GameRoutesMapController"
+      fx:controller="client.ui.game.GameRoutesMapController"
       prefHeight="1080" prefWidth="1920.0"
       stylesheets="/css/routesmapcss.css">
