+++ src/main/resources/views/layout_startup.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.Pane?>
 <Pane xmlns:fx="http://javafx.com/fxml"
       xmlns="http://javafx.com/javafx"
-      fx:controller="client.ui.MainMenuControllers.StartupController"
+      fx:controller="client.ui.mainmenu.StartupController"
       prefHeight="1080" prefWidth="3840"
       fx:id="rootPane" stylesheets="/css/defaultStyling.css">
