+++ src/main/resources/views/overlay_route_card_message.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.shape.Rectangle?>
 <StackPane xmlns:fx="http://javafx.com/fxml"
            xmlns="http://javafx.com/javafx"
-           fx:controller="client.ui.RouteCardMessageController"
+           fx:controller="client.ui.dialogs.RouteCardMessageController"
            prefHeight="600.0" prefWidth="1000.0"
            fx:id="rootPane">
